# docs/openapi/swaggerhub-export.json
{
  "paths": {
...
    "/settings": {
      "post": {
        "tags": [
          "visitor"
        ],
        "summary": "Get settings",
        "operationId": "settingsGet",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Getting settings for specific settings type",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "visitor"
        ],
        "summary": "Update settings",
        "operationId": "settingsUpdate",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Send updated settings for specific settings type",
          "required": true,
          "content": {
            "application/data": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "openapi": "3.0.0",
  "info": {
    "description": "Main API for StretchShop application",
    "version": "1.0.0-oas3",
    "title": "StretchShop main API",
    "contact": {
      "email": "support@stretchshop.app"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  }
}